pyproject.toml
[build-system]
requires = ["uv_build>=0.8.10,<0.9.0"]
build-backend = "uv_build"

[project]
name = "satterc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "jmarshrossney", email = "17361029+jmarshrossney@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.2",
    "pyproj>=3.7.2",
    "pyrealm>=2.0.0",
    "rioxarray>=0.22.0",
    "rothc-py",
    "sf-hamilton[visualization]>=1.89.0",
    "typer>=0.24.1",
    "xarray>=2026.2.0",
]

[project.scripts]
satterc = "satterc.cli:main"

[dependency-groups]
dev = [
    {include-group = "docs"},
    {include-group = "test"},
    "ruff>=0.15.4",
    "rust-just>=1.46.0",
]
docs = [
    "mkdocstrings-python>=2.0.3",
    "zensical>=0.0.24",
]
test = [
    "pytest>=9.0.2",
]

[tool.uv.sources]
rothc-py = { git = "https://github.com/satterc/rothc_py.git" }

